requirements.txt
absl-py==2.3.1
accelerate==1.10.1
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
audeer==2.2.2
audformat==1.3.2
audinterface==1.3.1
audiofile==1.5.1
audioread==3.0.1
audmath==1.4.2
audobject==0.7.12
audresample==1.3.4
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
cmake==4.1.0
cn2an==0.5.23
confluent-kafka==2.5.0
contourpy==1.3.0
cycler==0.10.0
cytoolz==1.0.1
datasets==4.1.1
decorator==5.2.1
diffusers==0.35.2
dill==0.4.0
distro==1.5.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
evaluate==0.4.6
faiss-cpu==1.13.0
faiss-gpu==1.7.2
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
graphviz==0.21
grpcio==1.76.0
hf-xet==1.1.10
httplib2==0.18.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==2.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
intervaltree==3.1.0
iotop==0.6
isc==2.0
iso3166==2.1.1
iso639-lang==2.6.3
jieba==0.42.1
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
k2==1.24.4.dev20241030+cuda12.1.torch2.5.1
kafka-python==2.0.2
kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=f935c2ac0be6dd1bf96a5852973dad6103c48e866cf33532459fb1336bea4bfe
kiwisolver==1.3.1
lazy_loader==0.4
lhotse==1.31.1
librosa==0.11.0
lilcom==1.8.1
llvmlite==0.43.0
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.9.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.2.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
olefile==0.46
omegaconf==2.3.0
openai-whisper==20250625
opensmile==2.6.0
overrides==7.7.0
oyaml==1.0
packaging==25.0
pandas==2.3.3
pesq==0.0.4
pillow==11.3.0
piper_phonemize==1.3.0
platformdirs==4.4.0
ply==3.11
pooch==1.8.2
proces==0.1.7
propcache==0.4.0
protobuf==6.33.0
psutil==7.1.0
pyarrow==21.0.0
pycparser==2.23
pycurl==7.43.0.6
pydub==0.25.1
pyparsing==2.4.7
pypinyin==0.55.0
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-dateutil==2.9.0.post0
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
rasterio==1.4.3
regex==2025.9.18
reportbug==7.10.3+deb11u1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.16.0
sonar-space==0.3.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
sympy==1.13.1
tableprint==0.9.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
toolz==1.0.0
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.5
vocos==0.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0

